API Remove code which is being migrated to a new TinyMCE module
